requirements-dev.txt
Pillow>=10.0.0
reportlab>=4.0.0
pypdf>=5.0.0
pytest>=8.0.0
